make project corrections
