+++ README.md
@@ -1,2 +1,43 @@
-# school_library
-School Library 
+# OOP school library
+Imagine that you are the librarian of OOP University, and you need a tool to record what books are in the library and who borrows them. The app that you will create will allow you to:
+
+- Add new students or teachers.
+- Add new books.
+- Save records of who borrowed a given book and when.
+
+## Getting Started
+-Download the repository as a zip file/ clone the repository using 
+git clone https://github.com/AbbyNyakara/school_library.git
+
+Create a branch and start working!
+
+## Authors
+
+👤 **Abby**
+
+- GitHub: [@AbbyNyakara](https://github.com/AbbyNyakara)
+- Twitter: [@AbigaelNyakara](https://twitter.com/AbbyNyakara)
+- LinkedIn: [Abigael Nyakara](https://linkedin.com/in/AbbyNyakara)
+
+
+## 🤝 Contributing
+
+Contributions, issues, and feature requests are welcome!
+
+Feel free to check the [issues page](../../issues/).
+
+## Show your support
+
+Give a ⭐️ if you like this project!
+
+## Acknowledgments
+
+- Coding Partner
+- Code Reviewer
+- etc
+
+## 📝 License
+
+This project is [MIT](https://spdx.org/licenses/MIT.html) licensed.
+
+
