+++ person.rb
@@ -1,8 +1,6 @@
-class Nameable
-  def correct_name
-    raise NotImplementedError
-  end
-end
+require_relative './nameable'
+require_relative './capitalize_decorator'
+require_relative './trimmer_decorator'
 
 class Person < Nameable
   attr_accessor :name, :age
@@ -33,35 +31,9 @@ class Person < Nameable
   private :of_age?
 end
 
-class NameableBaseDecorator < Nameable
-  attr_accessor :nameable
-
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < NameableBaseDecorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < NameableBaseDecorator
-  def correct_name
-    @nameable.correct_name
-    @nameable.correct_name[0...10] if @nameable.correct_name.length > 10
-  end
-end
-
 person = Person.new(22, 'maximilianus')
-person.correct_name
+p person.correct_name
 capitalized_person = CapitalizeDecorator.new(person)
-capitalized_person.correct_name
-capitalized_trimmed_person = TrimmerDecorator.new(capitalizedPerson)
-capitalized_trimmed_person.correct_name
+p capitalized_person.correct_name
+capitalized_trimmed_person = TrimmerDecorator.new(capitalized_person)
+p capitalized_trimmed_person.correct_name
